Move chat components to chat dir
